회전은 됨

+++ src/LabProject07-9-1/Player.h
@@ -57,7 +57,7 @@ public:
 	GameCamera* GetCamera() { return(myCamera); }
 	void SetCamera(GameCamera* pCamera) { myCamera = pCamera; }
 
-	void Move(ULONG nDirection, float fDistance, bool bVelocity = false);
+	void Move(DWORD nDirection, float fDistance, bool bVelocity = false);
 	void Move(const XMFLOAT3& xmf3Shift, bool bVelocity = false);
 	void Move(float fxOffset = 0.0f, float fyOffset = 0.0f, float fzOffset = 0.0f);
 	void Rotate(float x, float y, float z);
